\name{addSASexport}
\alias{addSASexport}
\title{Create code for SAS to export data to R}
\usage{
addSASexport(SAS2R = "", SAScache.directory = "",
  chunk.name = "unlabeled-SASnatch-chunk")
}
\arguments{
  \item{SAS2R}{character value, optional argument}

  \item{SAScache.directory}{character value, optional
  argument}

  \item{chunk.name}{character value, optional argument}
}
\description{
Create code for SAS to export data to R
}

