\name{addSASheader}
\alias{addSASheader}
\title{Create header for SAS code generated by SASnatch chunk}
\usage{
addSASheader(SAScache.directory = "",
  chunk.name = "unlabeled-SASnatch-chunk")
}
\arguments{
  \item{SAScache.directory}{character value, optional
  argument}

  \item{chunk.name}{character value, optional argument}
}
\description{
Create header for SAS code generated by SASnatch chunk
}
\examples{
addSASheader('~/courses/stat585/lab1/')
}

