\name{addSASimport}
\alias{addSASimport}
\title{Create code for SAS to import data from R}
\usage{
addSASimport(R2SAS = "", SAScache.directory = "",
  chunk.name = "unlabeled-SASnatch-chunk")
}
\arguments{
  \item{R2SAS}{character value, optional argument}

  \item{SAScache.directory}{character value, optional
  argument}

  \item{chunk.name}{character value, optional argument}
}
\description{
Create code for SAS to import data from R
}

