\name{catSnatch}
\alias{catSnatch}
\title{cat() printing for SASnatch objects}
\usage{
catSnatch(SASnatch.S4 = NULL, type = "rawcode")
}
\arguments{
  \item{SASnatch.S4}{SASnatch object, optional argument}

  \item{type}{character value, optional argument}
}
\description{
cat() printing for SASnatch objects
}

