\name{createSASfile}
\alias{createSASfile}
\title{Create the file containing the SAS data}
\usage{
createSASfile(code = "", SAScache.directory = "",
  chunk.name = "unlabeled-SASnatch-chunk")
}
\arguments{
  \item{code}{character value, optional argument}

  \item{SAScache.directory}{character value, optional
  argument}

  \item{chunk.name}{character value, optional argument}
}
\description{
Create the file containing the SAS data
}

