\name{helloSnatch}
\alias{helloSnatch}
\title{Greeting to be written after successfully loading package}
\usage{
helloSnatch(x = "")
}
\arguments{
  \item{x}{character value, optional argument}
}
\description{
Greeting to be written after successfully loading package
}
\examples{
helloSnatch('... hopefully')
}

