\name{makeSAScache}
\alias{makeSAScache}
\title{Create folder setup for SASnatch}
\usage{
makeSAScache(SASnatch.working.directory = "")

makeSAScache(SASnatch.working.directory = "")
}
\arguments{
  \item{SASnatch.working.directory}{character value,
  optional argument}

  \item{SASnatch.working.directory}{character value,
  optional argument}
}
\description{
Create folder setup for SASnatch

Create folder setup for SASnatch
}

