\name{printSASnatch}
\alias{printSASnatch}
\title{Print SASnatch results}
\usage{
printSASnatch(SASnatch = list(), type = "R", items = 0)
}
\arguments{
  \item{SASnatch.S4}{SASnatch object, optional argument}

  \item{type}{character value, optional argument}

  \item{items}{numeric vector, optional argument}
}
\description{
Print SASnatch results
}

