\name{read.SASnatch.object}
\alias{read.SASnatch.object}
\title{Create folder setup for SASnatch}
\usage{
read.SASnatch.object(chunk.name = "", SASresults.path = "",
  SAS2R.names = "", SAS2R.type = ".csv")
}
\arguments{
  \item{chunk.name}{character value, optional argument}

  \item{SASresults.path}{character value, optional
  argument}

  \item{SAS2R.names}{character vector, optional argument}

  \item{SAS2R.type}{character value, optional argument}
}
\description{
Create folder setup for SASnatch
}

