\name{read.SASnatch.results}
\alias{read.SASnatch.results}
\title{read results from SASnatch}
\usage{
read.SASnatch.results(chunk.name = "", SASresults.path = "",
  SAS2R.names = "", SAS2R.type = "csv")
}
\arguments{
  \item{chunk.name}{character value, optional argument}

  \item{SASresults.path}{character value, optional
  argument}

  \item{SAS2R.type}{character value, optional argument}

  \item{SAS2R.names}{character vector, optional argument}
}
\description{
read results from SASnatch
}

