\name{runSASnatch}
\alias{runSASnatch}
\title{Given the path to SAS, execute the SAS code created by SASnatch}
\usage{
runSASnatch(path_to_SAS.EXE = "", SAScache.directory = "",
  SASnatch.label = "")
}
\arguments{
  \item{path_to_SAS.EXE}{character value, optional
  argument}

  \item{SAScache.directory}{character value, optional
  argument}

  \item{SASnatch.label}{character value, optional argument}
}
\description{
Given the path to SAS, execute the SAS code created by
SASnatch
}

