\name{SASload.CSVdata}
\alias{SASload.CSVdata}
\title{Greeting to be written after successfully loading package}
\usage{
SASload.CSVdata(dsn = "")
}
\arguments{
  \item{dsn}{character vector, optional argument}

  \item{SASnatch}{character vector, optional argument}
}
\description{
Greeting to be written after successfully loading package
}
\examples{
SASload.CSVdata('d1, d2, d3')
}

