\name{SASnatch.before}
\alias{SASnatch.before}
\title{functions for knitr hooks}
\usage{
SASnatch.before(dsn, SASnatch.code, SASnatch.chunk_name = "SASnatch")
}
\arguments{
  \item{dsn}{character vector, optional argument}

  \item{SASnatch.code}{character value, optional argument}

  \item{SASnatch.chunk_name}{character value, optional
  argument}
}
\description{
functions for knitr hooks
}

