\name{SASnatch_hook}
\alias{SASnatch_hook}
\title{Hook function that guides SASnatch}
\usage{
SASnatch_hook(before, options, envir)
}
\arguments{
  \item{before}{logical, identifying before or after knitr
  chunk}

  \item{options}{list, options from knitr}

  \item{envir,}{options from R}
}
\description{
Hook function that guides SASnatch
}

