\name{snatchR2SAS}
\alias{snatchR2SAS}
\title{create the SAS code}
\usage{
snatchR2SAS(SAScache.directory = "", R2SAS = "", chunk.name = "")
}
\arguments{
  \item{SAScache.directory}{character value, optional
  argument}

  \item{R2SAS}{character value, optional argument}

  \item{chunk.name}{character value, optional argument}
}
\description{
create the SAS code
}

