\name{submitSAScode}
\alias{submitSAScode}
\title{Submit the SAS code generated by the SASnatch hook}
\usage{
submitSAScode(SAScache.directory = "", path_to_SAS.exe = "",
  SASnatch.chunk_name = "")
}
\arguments{
  \item{SASnatch.directory}{character value, optional
  argument}

  \item{path_to_SAS.exe}{character value, optional
  argument}

  \item{SASnatch.chunk_name}{character value, optional
  argument}
}
\description{
Submit the SAS code generated by the SASnatch hook
}

