\name{addSASfooter}
\alias{addSASfooter}
\title{Create footer for SAS code generated by SASnatch chunk}
\usage{
addSASfooter(working.directory = "",
  missing.chunk.name = "unlabeled-SASnatch-chunk")
}
\arguments{
  \item{working.directory}{character value, optional
  argument}

  \item{chunk.name}{character value, optional argument}
}
\description{
Create footer for SAS code generated by SASnatch chunk
}
\examples{
addSASfooter('~/courses/stat585/lab1/','chunk3')
}

