\name{addSASheader}
\alias{addSASheader}
\title{Create header for SAS code generated by SASnatch chunk}
\usage{
addSASheader(working.directory = "",
  missing.chunk.name = "unlabeled-SASnatch-chunk")
}
\arguments{
  \item{working.directory}{character value, optional
  argument}

  \item{missing.chunk.name}{character value, optional
  argument}
}
\description{
Create header for SAS code generated by SASnatch chunk
}
\examples{
addSASheader('~/courses/stat585/lab1/')
}

