\name{snatchSAS}
\alias{snatchSAS}
\title{call SAS without knitr}
\usage{
snatchSAS(dsn, code, S4name = "SnatchSAScode")
}
\arguments{
  \item{dsn}{character vector, optional argument}

  \item{code}{character value, optional argument}

  \item{S4name}{character value, optional argument}
}
\description{
call SAS without knitr
}

